Runs R commands on a cluster that supports SGE

Runs R commands on a cluster that supports SGE

gcluster.run2(..., command_list = NULL, opt.flags = "", max.jobs = 400,
  debug = FALSE, R = paste0(R.home(component = "bin"), "/R"),
  packages = NULL, jobs_title = NULL, job_names = NULL,
  collapse_results = FALSE, queue = NULL, memory = NULL, threads = NULL,
  io_saturation = NULL, queue_flag = "-q @{queue}",
  memory_flag = "-l mem_free=@{memory}G",
  threads_flag = "-pe threads @{threads}",
  io_saturation_flag = "-l io_saturation=@{io_saturation}",
  script = paste(Sys.getenv("ANALYSIS_HOME"), "common", "sgjob.sh", sep =
  "/"))

Arguments

...

R commands

command_list

list of strings with R commands

opt.flags

optional flags for qsub command

max.jobs

maximal number of simultaneously submitted jobs

debug

if 'TRUE', additional reports are printed

R

command that launches R

packages

packages to load for each command

jobs_title

title of job names. names would have the title followed by a serial number

job_names

vector with the names of the jobs

collapse_results

collapse return values of the jobs to a data frame. if not possible - would return the usual list.

memory

memory requirments (would be called using memory_flag)

threads

threads requirments (would be called using threads_flag)

io_saturation

io_saturation requirments (would be called using io_saturation_flag)

memory_flag

flag for memory requirment (formatted as in qq)

threads_flag

flag for threads requirment (formatted as in qq)

io_saturation_flag

flag for io_saturation requirment (formatted as in qq)

Value

if collapse_results is TRUE: data frame with the results of all jobs (rbinded). if collapse_results is FALSE returns the same as: gcluster.run

See also

gcluster.run

Examples