Align bisulfite converted reads to a reference genome using bowtie2

Align bisulfite converted reads to a reference genome using bowtie2

gpatterns.bissli2(r1_fastq, out_bam, genome_seq, bissli2_idx, r2_fastq = NULL,
  bissli2_bin = .gpatterns.bissli2_bin, bowtie2 = "bowtie2",
  samtools = "samtools", maxins = 1000, threads = 10,
  genome_type = "ct", tmp_dir = NULL, bissli2_params = "")

Arguments

r1_fastq

a vector of FASTQ files to align (read1)

out_bam

bam output file with aligned reads

genome_seq

A directory holding the reference genome, one FASTA file per chromosome

bissli2_idx

The basename of the index to be searched. The index is created using gpatterns.bissli2_build

r2_fastq

a vector of FASTQ files to align (read2)

bissli2_bin

path for bissli2-align

bowtie2

path for bowtie2

samtools

path for samtools

maxins

maximum fragment length (bowtie2 maxins argument)

threads

number of threads to use

genome_type

ct/ga/ct_ga. if 'ct' - Assume that the reads to be aligned underwent C->T conversion. If paired-end reads are aligned, then assume read 1 underwent C->T conversion while read 2 underwent G->A conversion. The --ct and --ga options are mutually exclusive. if 'ga' - Assume that the reads to be aligned underwent G->A conversion. If paired-end reads are aligned, then assume read 1 underwent G->A conversion while read 2 underwent C->T conversion. if 'ct_ga' reads would be aligned to both C->T and G->A, and the best match would be chosen.

tmp_dir

Directory for storing temporary files

bissli2_params

additional parameters to bissli2/bowtie2