Get coverage of intervals in multiple tracks

Get coverage of intervals in multiple tracks

gpatterns.intervals_coverage(tracks, intervals, iterator = NULL,
  file = NULL, intervals.set.out = NULL, pre_screen = TRUE,
  frag_cov = FALSE)

Arguments

tracks

methylation tracks

intervals

genomic scope for which the function is applied

iterator

see iterator in gextract. if NULL iterator would be set to CpGs

file

save output to file (only in non tidy mode, would not filer by variance)

intervals.set.out

save output big intervals set (only in tidy mode, would not filter by variance)

pre_screen

screen for covered intervals before calculating the total coverage (for large intervals set)

frag_cov

use fragment coverage (if exists)

Value

intervals set with 'sum' column containing the total coverage (sum) in all the tracks