Merge tracks

Creates a track called new_track that is the sum of the methylation calls in tracks. In addition, if merge_tidy_cpgs is TRUE, creates a union of tidy_cpgs of tracks and generates pattern frequency and pattern coverage for the combined tidy_cpgs. If patterns attributes are present, they will be combined and calculated for the new track.

gpatterns.merge_tracks(tracks, new_track, description,
  intervals = gintervals.all(), iterator = NULL, add_var = FALSE,
  merge_tidy_cpgs = TRUE, ...)

Arguments

tracks

tracks to merge

new_track

new track name

description

new track description

intervals

intervals scope. default is all the genome (gintrevals.all())

iterator

new track iterator. if left NULL, the iterator would be the union of all the covered CpGs of tracks

add_var

calculate variance (of tracks '.avg'). Will be created in new_track.var

merge_tidy_cpgs

merge also tidy cpgs and create pattern frequency and pattern coverage track. Does not work for tracks without the single molecule data (e.g. arrays, processed data)

...

additional parameters for gpatterns.import_from_tidy_cpgs

Value

name of the new_track