Plot a heatmap of average methylation clustering

Plot a heatmap of average methylation clustering

gpatterns.plot_clustering(avgs, rows_clust_method = "kmeans", K = NULL,
  cluster_cols = TRUE, show_row_clust = TRUE, annotation = NULL,
  annotation_colors = NULL, annotation_col = NULL, show_rownames = FALSE,
  show_colnames = FALSE, cluster_gaps = FALSE, color_pal = .blue_red_pal,
  fontsize = 8, fontsize_row = fontsize, fontsize_col = fontsize,
  main = NULL, device = "png", fig_ofn = NULL, width = 2100,
  height = 2000, res = 300, plot = TRUE, motif_enrich = NULL,
  cluster_motifs = TRUE, qval_thresh = 0.05, show_sig = TRUE,
  motif_fontsize_col = fontsize * 0.8, motif_fontsize_sig = fontsize * 1.3,
  ...)

Arguments

avgs

output of gpatterns.cluster_avg_meth

cluster_cols

boolean values determining if columns should be clustered or hclust object.

annotation

deprecated parameter that currently sets the annotation_col if it is missing

annotation_colors

list for specifying annotation_row and annotation_col track colors manually. It is possible to define the colors for only some of the features. Check examples for details.

annotation_col

similar to annotation_row, but for columns.

show_rownames

boolean specifying if column names are be shown.

show_colnames

boolean specifying if column names are be shown.

fontsize

base fontsize for the plot

fontsize_row

fontsize for rownames (Default: fontsize)

fontsize_col

fontsize for colnames (Default: fontsize)

main

the title of the plot

fig_ofn

output filename of the figure. if NULL would plot to current device

width

width ('png' parameter)

height

height ('png' parameter)

plot

if FALSE, would only return the clustering

...

other pheatmap1 parameters