Define putative enhancer regions from chip-seq tracks

Define putative enhancer regions from chip-seq tracks

gpatterns.putative_enhancers(chip_tracks, quant_thresh = 0.999,
  normalize = NULL, min_tss_dist = 2000, logical_gate = "|")

Arguments

chip_tracks

names of chip-seq tracks

quant_thresh

quantile of chip signal that would be considered as peak

normalize

normalize all the regions to have the same size. if NULL no normalization would be done.

min_tss_dist

minimal (absolute) distance from TSS

logical_gate

logical function to combine the different tracks, e.g. '|' or '&'

Value

intervals set with putative enhancers