Plots smoothScatter of average methylation of two tracks

Plots smoothScatter of average methylation of two tracks

gpatterns.smoothScatter(samples, avgs = NULL, fig_ofn = NULL,
  sample_names = NULL, width = 300, height = 300,
  color.pal = .smooth_scatter_pal2, device = "png", title_text = NULL,
  add_n = TRUE, min_cov = NULL, min_samples = 2, iterator = NULL, ...)

Arguments

samples

a vector of 2 sample names as present in 'samp' filed of avgs, or track names if avgs is NULL

avgs

pre-computed intervals set with average methylation, output of gpatterns.get_avg_meth. if NULL, gpatterns.get_avg_meth(tracks=samples, ...) would be called. Note that you using this option together with additional parameters to smoothScatter would produce warnings.

fig_ofn

figure filename. if NULL figure would be plotted on the current device

sample_names

names for the two tracks to plot as the axis labels. if NULL the sample names would be used.

width

plot width

height

plot height

color.pal

color pallete

device

function to open a device if fig_ofn is not NULL. default: 'png'

title_text

title text

add_n

add 'n = number_of_observations' to the title

min_cov

minimal coverage for iterator interval.

...

other parameters for smoothScatter / gpatterns.get_avg_meth.

Value

None

Examples