wrapper around pheatmap

wrapper around pheatmap

pheatmap1(pmat, annotation = NULL, annotation_colors = NULL,
  annotation_col = NULL, annotation_row = NULL, id_column = TRUE,
  border_color = NA, clustering_callback = function(hc, ...) {
  dendsort::dendsort(hc) }, clustering_distance_rows = "euclidean",
  clustering_distance_cols = "euclidean", min_vals_row = 1,
  min_vals_col = 1, na_col = "gray", ...)

Arguments

pmat

data frame to plot as matrix. can have an additional id column (if id_column == TRUE)

annotation

data frame with 'samp' field, and additional annotations

annotation_colors

tidy data frame with 3 fields: 'type' - name of the annotation, e.g. 'tissue', and pairs of 'variable' and 'color', e.g. variable='tumor', color='green'

annotation_col

character vector with column names from `annotation` to use to annotate columns

annotation_row

character vector with column names from `annotation` to use to annotate rows

id_column

TRUE if pmat's first column contains row ids

border_color

color of cell borders on heatmap, use NA if no border should be drawn.

clustering_callback

callback function to modify the clustering. Is called with two parameters: original hclust object and the matrix used for clustering. Must return a hclust object.

clustering_distance_rows

distance measure used in clustering rows. Possible values are "correlation" for Pearson correlation and all the distances supported by dist, such as "euclidean", etc. If the value is none of the above it is assumed that a distance matrix is provided.

clustering_distance_cols

distance measure used in clustering columns. Possible values the same as for clustering_distance_rows.

na_col

specify the color of the NA cell in the matrix.

...

graphical parameters for the text used in plot. Parameters passed to grid.text, see gpar.